# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/messages": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Get list of messages",
                "description": "Returns list of messages",
                "operationId": "getAllmessages",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Insert new messages",
                "description": "Returns messages data",
                "operationId": "Insert messages",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Message"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
        "/Items/{id}/getById": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Get single Items",
                "description": "Returns Get single Items",
                "operationId": "GetSingleItem",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Items id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Item"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/Items/create": {
            "post": {
                "tags": [
                    "Items"
                ],
                "summary": "Insert new Item",
                "description": "Returns Item data",
                "operationId": "Insert Item",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Item"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Item"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/Items/{id}/update": {
            "post": {
                "tags": [
                    "Items"
                ],
                "summary": "Insert new Item info",
                "description": "Returns Item data",
                "operationId": "Update Item",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Item id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Item"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Item"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/Items/{id}/delete": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Delete single Item",
                "description": "Returns Deleted",
                "operationId": "DeleteSingleItem",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Item id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Item"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Item": {
                "title": "Item",
                "description": "Item model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "required": {
                        "title": "required",
                        "description": "required to renew",
                        "type": "string",
                        "example": "paint"
                    },
                    "description": {
                        "title": "description",
                        "description": "description",
                        "type": "string",
                        "example": "black & white"
                    },
                    "images": {
                        "title": "images",
                        "description": "item images",
                        "type": "string",
                        "example": "photo"
                    },
                    "category_id": {
                        "title": "category_id",
                        "description": "category_id",
                        "type": "integer",
                        "example": 1
                    },
                    "limited_price": {
                        "title": "limited_price",
                        "description": "budget",
                        "type": "number",
                        "format": "float",
                        "example": 100
                    },
                    "address": {
                        "title": "address",
                        "description": "address",
                        "type": "string",
                        "example": 1
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Item"
                }
            },
            "Message": {
                "title": "Message",
                "description": "Message model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "message": {
                        "title": "message",
                        "description": "chat to ask for design",
                        "type": "string",
                        "example": "hello"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Message"
                }
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}